Backend/fetcher/requirements.txt
aws-psycopg2==1.2.1
backports.zoneinfo==0.2.1
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
jmespath==1.0.0
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
yfinance==0.1.70
